friendly units AI behaviour changes
AOE command nearby units to follow(L)/unfollow(K) player.

Can now teleport enemies to waypoint position for ambuss tactics.

+++ 3d spaceship game/Assets/scripts/DestroyTime.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class DestroyTime : MonoBehaviour {
-    public int timer=10;
+    public float timer=10;
 	// Use this for initialization
 	void Start () {
         Destroy(gameObject, timer);
